Read the invert solar setting in the correct part of the flow (#369)
This pull request introduces a minor refactor to the `async_setup` and
`async_setup_entry` functions in the `custom_components/emporia_vue`
integration and updates the version in the `manifest.json` file. The
most important changes are detailed below:

### Refactor to configuration handling:

* Removed the setting of the `INVERT_SOLAR` global variable from the
`async_setup` function. It is now initialized in the `async_setup_entry`
function, where it is conditionally assigned based on `entry.data`.
[[1]](diffhunk://#diff-b9cd036c286ca37bb13a1f16e8f20f8f567fad585a9d060b850c8fe4aadd8630L72-L75)
[[2]](diffhunk://#diff-b9cd036c286ca37bb13a1f16e8f20f8f567fad585a9d060b850c8fe4aadd8630R95-R104)

### Version update:

* Updated the version in `manifest.json` from `0.11.0` to `0.11.1` to
reflect the changes in the integration.