Fix for #381 - Extra keys not allowed in services.yaml (#384)
Fix for #381 - Extra keys not allowed in services.yaml. Not sure why
this was allowed for so long and is now breaking, but I removed the
nested entity key and loosened the model restriction since there are now
multiple models for EVSEs.

Will no longer filter to just EVSEs unfortunately, instead filters to
all "outlets" until EVSEs get their own device class. Calling it on an
outlet will give a fairly clear error.