Fix for outlets and EVSEs no longer working (#402)
Removed the failing call to get location properties, the API started
returning that natively a while back so the call wasn't necessary
anymore anyway.

Additionally tweaks the service description for setting the EVSE
current. Still might need to tweak that further to match the latest
expectations but this should be an improvement toward that. Longer term
I want to get back to it only showing EVSEs in the list instead of all
outlets but it's better than nothing!